% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{auc}
\alias{auc}
\title{Calculate AUC}
\usage{
auc(probabilities, truth, negative, positive)
}
\arguments{
\item{probabilities}{[\code{numeric}]\cr
The predicted probabilities for the positive class as a
numeric vector.}

\item{truth}{[\code{factor} or \code{numeric}]\cr
Vector of the true class.}

\item{negative}{[\code{character}]\cr
The name of the negative class.}

\item{positive}{[\code{character}]\cr
The name of the positive class}
}
\value{
The Area under the Receiver Operating Curve as single \code{numeric}.
}
\description{
Computes the Area under the Receiver Operating Curve.
}
