% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{bac}
\alias{bac}
\title{Calculate BAC}
\usage{
bac(response, truth, negative, positive)
}
\arguments{
\item{response}{[\code{factor}]\cr
Vector of the predicted class.}

\item{truth}{[\code{factor} or \code{numeric}]\cr
Vector of the true class.}

\item{negative}{[\code{character}]\cr
The name of the negative class.}

\item{positive}{[\code{character}]\cr
The name of the positive class}
}
\value{
The balanced accuracy as single \code{numeric}.
}
\description{
Computes the balanced accuracy.
}

