% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{mbmdr}
\alias{mbmdr}
\title{Run the MB-MDR algorithm on the data}
\usage{
mbmdr(data, order = 2L, alpha = 0.1, max.results = 100)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
The data.}

\item{order}{[\code{int}]\cr
The interaction depth, i.e. how many features are in one
model.}

\item{alpha}{[\code{double}]\cr
Significance threshold for classifying a genotype combination
as (H)igh or (L)ow risk.}

\item{max.results}{[\code{int}]\cr
Report only this amount of top models.}
}
\value{
A \code{list} object of class \code{mbmdr} containing the top models
and their respective test statistic, the HLO tables of the top models, the
model tables of the top models and the function arguments \code{order},
\code{alpha} and \code{max.results}.
}
\description{
Successively calculate the test statistics of all possible models, i.e.
feature combinations in the data.
}

