% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdrc.R
\name{mbmdrc}
\alias{mbmdrc}
\title{MB-MDR based classifier}
\usage{
mbmdrc(formula, data, order = 2, min_cell_size = 10, alpha = 0.1,
  max_results = 100, top_results = 10, folds, cv_loss,
  dependent_variable_name, num_threads, verbose, log_file, ...)
}
\arguments{
\item{formula}{[\code{formula}]\cr
A formula of the form LHS ~ RHS, where LHS is the dependent
variable and RHS are the features to enter the MB-MDR
 genotype classification.}

\item{data}{[\code{data.frame}]\cr
The data object.}

\item{order}{[\code{int}]\cr
Single integer specifying the interaction depth used
in the MB-MDR algorithm.}

\item{min_cell_size}{[\code{int}]\cr
Single integer specifying the minimum number of
observation in a genotype combination to be statistically
relevant.}

\item{alpha}{[\code{double}]\cr
Single numeric as significance level used during HLO
classification of genotype combinations.}

\item{max_results}{[\code{int}]\cr
Single integer specifying the number of top models to
report.}

\item{top_results}{[\code{int}]\cr
Single integer specifying how many models shall be used
for prediction. If \code{folds} and \code{cv.loss} are
set, this value specifies the upper limit of top results
to assess in CV.}

\item{folds}{[\code{int}]\cr
Single interger specifying how many folds for internal
cross validation to find optimal \code{top_results}
should be used.}

\item{cv_loss}{[\code{character}]\cr
One of \code{auc} or \code{bac}, specifying which loss
should be used to find optimal \code{top_results}.}

\item{dependent_variable_name}{[\code{string}]\cr
Name of dependent variable, needed if no formula given.}

\item{num_threads}{[\code{int}]\cr
Number of threads. Default is number of CPUs available.}

\item{verbose}{[\code{int}]\cr
Level of verbosity. Default is level 1 giving some basic
information about progress. Level 0 will switch off any
output. Levels 2 and 3 are for debugging purposes.}

\item{log_file}{[\code{string}]\cr
File path for a log file. If empty (default) logging is
sent to console.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A S3 object of class \code{mbmdrc}.
}
\description{
Use the MB-MDR HLO classification of top models to predict the disease risk.
}
\details{
If the data type of the dependent variable is a factor, classification mode is started automatically. Otherwise MB-MDR will run in regression mode.
In classification mode, the first factor level is assumed to code for the negative class.
}
