% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{NextModel}
\alias{NextModel}
\title{Get the next successive model}
\usage{
NextModel(model, k, order, j = order)
}
\arguments{
\item{model}{[\code{integer}]\cr
The current model.}

\item{k}{[\code{int}]\cr
Total number of features.}

\item{order}{[\code{int}]\cr
Interaction order.}

\item{j}{[\code{int}]\cr
The model position to increase. If called, this is usually the
interaction order.}
}
\value{
A \code{integer} vector of features. If the last model is reached,
\code{FALSE} is returned.
}
\description{
This function returns the next feature combination, i.e. model, given the
current model, the order and the total number of features.
}

