% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdrc.R
\name{predict.mbmdrc}
\alias{predict.mbmdrc}
\title{MB-MDR classifier prediction}
\usage{
\method{predict}{mbmdrc}(object, newdata, type = "response", top_results,
  num_threads, verbose, ...)
}
\arguments{
\item{object}{[\code{mbmdr}]\cr
MB-MDR models and HLO tables as output from \code{\link{mbmdr}}.}

\item{newdata}{[\code{newdata}]\cr
New data to predict class status for.}

\item{type}{[\code{string}]\cr
Type of prediction. One of \code{response}, \code{prob},
\code{score} or \code{scoreprob}. See details.
Default is \code{response}.}

\item{top_results}{[\code{int}]\cr
How many models are used for prediction.}

\item{...}{Further arguments passed to or from other methods.}

\item{o.as.na}{[\code{bool}]\cr
Indicates how to handle samples in O classified genotype
combinations. See details.
Default is \code{TRUE}.}
}
\value{
A \code{data.table} object with an ID column and the prediction value.
}
\description{
Prediction with new data and saved MB-MDR classifier object.
}
\details{
For \code{type='response'} (the default), the predicted classes, for
\code{type='prob'} the predicted case probabilities, for \code{type='score'}
the predicted risk scores and for \code{type='scoreprob'} the predicted risk
scores transformed to the [0, 1] interval are returned.

For \code{type='score'} and \code{type='scoreprob'} genotypes classified as
H contribute +1, as L contribute -1 and as O contribute 0 to the score.

If a genotype combination is classified as O by MB-MDR, the case probability
is not significantly different from 0.5. On the other hand, there might have
been just too few observations in the training data so that \code{NA} might
be more reasonable as contribution to \code{response} and \code{prob} type
predictions.
}

