% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{Step1}
\alias{Step1}
\title{Calculate HLO matrix}
\usage{
Step1(data, model, mu, alpha)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
The data.}

\item{model}{[\code{integer}]\cr
The feature combination to consider.}

\item{mu}{[\code{double}]\cr
The overall case to control ratio.}

\item{alpha}{[\code{double}]\cr
The significane level.}
}
\value{
A \code{data.table} object. Each row is a genotype combination of the
model. The columns contain the counts of cases and controls, the case ratio in
the genotype combination, the case ratio in all other genotype combinations,
the chi-square statistic, the corresponding p value and a class label.
}
\description{
For each model calculate the case ratio for each genotype combination
and test if this ratio is significantly different to the case ratio
in all other genotype combinations.
}

