% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{Step2}
\alias{Step2}
\title{Assess significance of model}
\usage{
Step2(counts, mu)
}
\arguments{
\item{counts}{[\code{data.table}]\cr
Output of \code{\link{Step1}}.}

\item{mu}{[\code{double}]\cr
The overall case to control ratio.#'}
}
\value{
A \code{data.table} object, containing the counts of cases and controls
in each class of H, L or O, their respective ratios in and not in the class and
the chi-square test statistic.
}
\description{
After classifying each genotype combination of a model as (H)igh, (L)ow or
uninformative (O), assess the significance of the model.
}

