% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{tn}
\alias{tn}
\title{Calculate TN}
\usage{
tn(response, truth, negative)
}
\arguments{
\item{response}{[\code{factor}]\cr
Vector of the predicted class.}

\item{truth}{[\code{factor} or \code{numeric}]\cr
Vector of the true class.}

\item{negative}{[\code{character}]\cr
The name of the negative class.}
}
\value{
The number of true negatives as single \code{numeric}.
}
\description{
Computes the number of true negatives
}
