% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{tp}
\alias{tp}
\title{Calculate TP}
\usage{
tp(response, truth, positive)
}
\arguments{
\item{response}{[\code{factor}]\cr
Vector of the predicted class.}

\item{truth}{[\code{factor} or \code{numeric}]\cr
Vector of the true class.}

\item{positive}{[\code{character}]\cr
The name of the positive class.}
}
\value{
The number of true positives as single \code{numeric}.
}
\description{
Computes the number of true positives.
}

