% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdrc.R
\name{mbmdrc}
\alias{mbmdrc}
\title{MB-MDR based classifier}
\usage{
mbmdrc(formula, data, order = 2, alpha = 0.1, max.results = 100,
  top.results = 10, folds, cv.loss)
}
\arguments{
\item{formula}{[\code{formula}]\cr
A formula of the form LHS ~ RHS, where LHS is the disease
status and RHS are the features to enter the MB-MDR
genotype classification.}

\item{data}{[\code{data.frame}]\cr
The data object.}

\item{order}{[\code{int}]\cr
Single integer specifying the interaction depth used
in the MB-MDR algorithm.}

\item{alpha}{[\code{double}]\cr
Single numeric as significance level used during HLO
classification of genotype combinations.}

\item{max.results}{[\code{int}]\cr
Single integer specifying the number of top models to
report.}

\item{top.results}{[\code{int} or \code{integer}]\cr
Single integer specifying how many models shall be used
for prediction. If \code{folds} and \code{cv.loss} are
set, this value specifies the upper limit of top results
to assess in CV, or, if a \code{integer} vector the top
result values to assess in CV.}

\item{folds}{[\code{int}]\cr
Single interger specifying how many folds for internal
cross validation to find optimal \code{top.results}
should be used.}

\item{cv.loss}{[\code{character}]\cr
One of \code{auc} or \code{bac}, specifying which loss
should be used to find optimal \code{top.results}.}
}
\value{
A S3 object of class \code{mbmdrc}.
}
\description{
Use the MB-MDR HLO classification of top models to predict the disease risk.
}

