% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdrc.R
\name{mbmdrc}
\alias{mbmdrc}
\title{MB-MDR based classifier}
\usage{
mbmdrc(formula, data, order = 2, min.cell.size = 10, alpha = 0.1,
  adjustment = "NONE", max.results = 1000, top.results = 1000, folds,
  cv.loss, o.as.na, dependent.variable.name, verbose, ...)
}
\arguments{
\item{formula}{[\code{formula}]\cr
A formula of the form LHS ~ RHS, where LHS is the dependent
variable and RHS are the features to enter the MB-MDR
 genotype classification.}

\item{data}{[\code{data.frame}]\cr
The data object.}

\item{order}{[\code{int} or \code{integer}]\cr
Single integer specifying the interaction depth used
in the MB-MDR algorithm. Possible options are
1 and 2.}

\item{min.cell.size}{[\code{int}]\cr
Single integer specifying the minimum number of
observation in a genotype combination to be statistically
relevant.}

\item{alpha}{[\code{double}]\cr
Single numeric as significance level used during HLO
classification of genotype combinations.}

\item{adjustment}{[\code{string}]\cr
Adjust method to be used.
"CODOMINANT", "ADDITIVE" or "NONE" (default).}

\item{max.results}{[\code{int}]\cr
Single integer specifying the number of top models to
report.}

\item{top.results}{[\code{int}]\cr
Single integer specifying how many models shall be used
for prediction. If \code{folds} and \code{cv.loss} are
set, this value specifies the upper limit of top results
to assess in CV.}

\item{folds}{[\code{int}]\cr
Single interger specifying how many folds for internal
cross validation to find optimal \code{top.results}
should be used.}

\item{cv.loss}{[\code{character}]\cr
One of \code{auc} or \code{bac}, specifying which loss
should be used to find optimal \code{top.results}.}

\item{o.as.na}{[\code{bool}]\cr
Encode non informative cells with NA or with 0.5.}

\item{dependent.variable.name}{[\code{string}]\cr
Name of dependent variable, needed if no formula given.}

\item{verbose}{[\code{int}]\cr
Level of verbosity. Default is level 1 giving some basic
information about progress. Level 0 will switch off any
output.}

\item{...}{Arguments passed from and to other functions.}
}
\value{
A S3 object of class \code{mbmdrc}.
}
\description{
Use the MB-MDR HLO classification of top models to predict the disease risk.
}
\details{
If the data type of the dependent variable is a factor, classification mode is started automatically. Otherwise MB-MDR will run in regression mode.
In classification mode, the first factor level is assumed to code for the negative class.
}
